Make librrd.c work with RRDD's inotify-based discovery mechanism
(1) This commit changes behaviour of the library to accommodate RRDD's
new plugin discovery mechanism:

* The library writes a file of static size and doesn't resize it. The
  size depends on RRD_MAX_SOURCES and no

* The library immediately writes to the file after it opens it such that
  it contains a valid header, which is picked up by RRDD.

* The library removes the file it has written when a plugin is closed.
  This is picked up by RRDD.

(2) Furthermore, the commit contains these general improvements:

* Linking the shared library requires zlib.

* Check for memory leaks using valgrind as part of CI.

* integrate new upstream release of JSON lib parson

(3) The interface for clients changed: the sample() function now
receives a pointer provided by the client. This makes it easier to write
a sample function that is shared by multiple data sources. See the
rrdtest.c for an example.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>