CA-223651 fix meta data for VM and SR owner (#11)
This commit fixes the owner meta data for VMs and SRs. The string
representing them need to include the UUID. The testing code rrdclient.c
was changed to now use RRD_VM in order to test this as part of the
integration test. Wait 2 seconds to give the client a chance to write data before trying
to read it.

Use a static initialiser for owner[] buffer. This is not strictly
necessary as snprintf() is guaranteed to terminate it with a null byte.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>