librrd: Get timestamp at a higher precision
Changes the usage of `time(2)` to `gettimeofday(2)`, gaining greater
precision of the timestamp transmitted to the server (now sent as a
double instead of uint64). Due to different signatures of these
functions, the second parameter to `rrd_sample` was deprecated and is
now simply ignored. No known users of this plugin set it to anything
other than NULL so this is not disruptive.

This follows the change in the RRD protocol upstream.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>