CA-414341: PVS RRD graph does not show
The PR https://github.com/xapi-project/xen-api/pull/6086 changed the
timestamp of rrd data from `int64` to `double`. As a result, the timestamp
in RRD messages sent from `pvsproxy` to `xcp-rrdd` is inconsistent between
the two formats, causing parsing failures and resulting in missing data.

Another PR https://github.com/xapi-project/rrd-client-lib/pull/14
attempted to address the issue by introducing a new get_timestamp function
that generates the timestamp as a double. However, applying htonll to this
value unintentionally causes implicit conversion from the double-precision
to int64_t, which preserves only the integer part and discards the
fraction. Since xcp-rrdd requires the timestamp in IEEE 754 double format,
this int64_t value could not be parsed correctly.

The solution is to returns the timestamp as an `uint64_t` containing the
IEEE 754 bit pattern of the double value.

Signed-off-by: Bengang Yuan <bengang.yuan@cloud.com>